p,r,t=map(int,input().split()) k=1+(r/100) i=(p*(k**t)) print("{:.2f}".format(i))

+++ Compound_interest.py
@@ -1,4 +1,4 @@
-p,r,t=map(float,input().split())
+p,r,t=map(int,input().split())
 k=1+(r/100)
 i=(p*(k**t))
 print("{:.2f}".format(i))
